Finish metrics
