+++ function.go
@@ -3,9 +3,12 @@ package dataqual
 import (
 	"context"
 	"fmt"
+	"io"
 
 	"github.com/pkg/errors"
+	"github.com/tetratelabs/wazero"
 	"github.com/tetratelabs/wazero/api"
+	"github.com/tetratelabs/wazero/imports/wasi_snapshot_preview1"
 )
 
 type function struct {
@@ -124,4 +127,26 @@ func createFunction(wasmBytes []byte) (*function, error) {
 		alloc:   alloc,
 		dealloc: dealloc,
 	}, nil
+}
+
+func createWASMInstance(wasmBytes []byte) (api.Module, error) {
+	if len(wasmBytes) == 0 {
+		return nil, errors.New("wasm data is empty")
+	}
+
+	ctx := context.Background()
+	r := wazero.NewRuntime(ctx)
+
+	wasi_snapshot_preview1.MustInstantiate(ctx, r)
+	cfg := wazero.NewModuleConfig().
+		WithStderr(io.Discard).
+		WithStdout(io.Discard).
+		WithStartFunctions("") // We don't need _start() to be called for our purposes
+
+	mod, err := r.InstantiateWithConfig(ctx, wasmBytes, cfg)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to instantiate wasm module")
+	}
+
+	return mod, nil
 }
